requirements.txt
Django
django-taggit
django-admin-chart
django-admin-charts
django-admin-tools
django-admin-tools-stats
django-hitcount
django-jazzmin
pillow
dj-database-url
python-dotenv
psycopg2-binary